code optimization changes remove unused method,charctres

+++ lib/models/network_model/order_model/order_request_model.dart
@@ -52,7 +52,7 @@ class PlaceOrderRequestModel {
   String? zipCode;
   bool? anonymous;
   int? donation;
-  int? gratuity;
+  double? gratuity;
   int? ccChargesIncluded;
   int? corporateDonation;
   num? corporateDonationBeforeCcCharges;
